+++ src/types/config.ts
@@ -1,3 +1,5 @@
+import { CipherGCMTypes } from 'crypto';
+
 export interface ActionInputs {
   icredible_activation_code: string;
   icredible_encryption_password: string;
@@ -16,10 +18,13 @@ export interface ApiConfig {
 }
 
 export interface CryptoConfig {
-  algorithm: string;
-  keyDerivation: string;
-  compressionLevel: number;
-  hashAlgorithm: string;
+  algorithm: CipherGCMTypes;
+  digest: string;
+  saltLength: number;
+  ivLength: number;
+  keyLength: number;
+  iterations: number;
+  authTagLength: number;
 }
 
 export interface FileConfig {